# client-api/openapi-spec.yml
openapi: '3.0.0'

info:
  title: 'Client API'
  version: '1.0.0-alpha'
  description: Endpoint for user to communicate with the service

servers:
  - url: 'http://localhost:5000'

paths:
  /api/users/register:
    post:
      summary: 'Register a new user'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        '201':
          description: 'User registered successfully'
        '400':
          description: 'Invalid data provided'

  /api/users/auth:
    post:
      summary: 'Auth user'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        '200':
          description: 'Auth success, returning auth token'
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          description: 'Invalid credentials'

  /api/users/update:
    post:
      summary: 'Update user data'
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                last_name:
                  type: string
                birthdate:
                  type: string
                  format: date
                email:
                  type: string
                  format: email
                phone:
                  type: string
                  format: phone
              required: []
      responses:
        '200':
          description: 'Data updated successfully'
        '400':
          description: 'Invalid data provided'
        '403':
          description: 'Invalid auth token'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT